Move tree state management to Sidebar component
