+++ html-report/src/components/sidebar/keys.ts
@@ -1,0 +1,4 @@
+import {InjectionKey} from "vue";
+import {RootUi} from "./TreeState.ts";
+
+export const rootStoreKey = Symbol() as InjectionKey<RootUi>
