+++ html-report/src/main.ts
@@ -1,4 +1,4 @@
-import {createApp, Reactive, reactive} from 'vue'
+import {createApp} from 'vue'
 import {createI18n} from 'vue-i18n'
 
 import './style.css'
@@ -6,108 +6,12 @@ import App from './App.vue'
 import TestExecution from './TestExecution'
 import VueEasyLightbox from "vue-easy-lightbox";
 
-const rootProps = {
+const app = createApp(App, {
   executions: globalThis.testExecutions
     .map(it => new TestExecution(it)),
-};
+});
 
-export type NodeUi = {
-    collapsed: boolean
-}
-
-export type RootUi = {
-    nodes: Record<string, NodeUi>
-    showAborted: boolean
-    showFailed: boolean
-    showSkipped: boolean
-    showSuccessful: boolean
-
-    collapseAll: () => void
-    expandAll: () => void
-    toggleShowAborted: () => void
-    toggleShowFailed: () => void
-    toggleShowSkipped: () => void
-    toggleShowSuccessful: () => void
-    toggleNode: (id: string) => void
-
-    isVisible: (statuses: string[]) => boolean
-}
-
-export const rootStore: Reactive<RootUi> = reactive<RootUi>({
-        nodes: rootProps.executions
-            .reduce((prev, execution) => {
-                return {
-                    ...prev,
-                    [execution.id]: {
-                        collapsed: false,
-                    },
-                    ...execution.nodesWithChildren().reduce((prev, node) => {
-                        const statuses = execution.nodeStatuses(node);
-                        let initiallyCollapsed = execution.parents(node).length > 1 && statuses.indexOf('FAILED') == -1;
-                        return {
-                            ...prev,
-                            [node.id]: {
-                                collapsed: initiallyCollapsed,
-                            }
-                        }
-                    }, {})
-                }
-            }, {} as Record<string, NodeUi>),
-        showAborted: true,
-        showFailed: true,
-        showSkipped: true,
-        showSuccessful: true,
-
-        toggleShowAborted() {
-            this.showAborted = !this.showAborted;
-        },
-
-        toggleShowFailed() {
-            this.showFailed = !this.showFailed;
-        },
-
-        toggleShowSuccessful() {
-            this.showSuccessful = !this.showSuccessful;
-        },
-
-        toggleShowSkipped() {
-            this.showSkipped = !this.showSkipped;
-        },
-
-        collapseAll() {
-            Object.keys(this.nodes).forEach((key) => {
-                this.nodes[key].collapsed = true;
-            });
-        },
-        expandAll() {
-            Object.keys(this.nodes).forEach((key) => {
-                this.nodes[key].collapsed = false;
-            });
-        },
-        toggleNode(id: string) {
-            this.nodes[id].collapsed = !this.nodes[id].collapsed;
-        },
-
-        isVisible(statuses: string[]): boolean {
-            return statuses.length == 0 || statuses.filter(status => {
-                switch (status) {
-                    case 'SUCCESSFUL':
-                        return this.showSuccessful;
-                    case 'FAILED':
-                        return this.showFailed;
-                    case 'SKIPPED':
-                        return this.showSkipped;
-                    case 'ABORTED':
-                        return this.showAborted;
-                }
-            }).length > 0;
-        }
-
-    }
-)
-const app = createApp(App, rootProps);
-
-const i18n = createI18n({
+app.use(createI18n({
   messages: {
     en: {
       executionSummary: {
@@ -136,8 +40,8 @@ const i18n = createI18n({
       }
     }
   }
-});
+}));
 
-app.use(i18n);
 app.use(VueEasyLightbox);
+
 app.mount('#app');
